// web/src/api/generated/model/changedTransactionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Couplefins
 * OpenAPI spec version: 0.1.0
 */
import type { FieldDiffResponse } from './fieldDiffResponse';
import type { PreviewTransactionResponse } from './previewTransactionResponse';

export interface ChangedTransactionResponse {
  existing_id: string;
  incoming: PreviewTransactionResponse;
  existing: PreviewTransactionResponse;
  diffs: FieldDiffResponse[];
}